moved helper to separate package
