+++ internal/archive_test.go
@@ -2,9 +2,7 @@ package internal
 
 import (
 	"errors"
-	"fmt"
 	"log/slog"
-	"path/filepath"
 	"regexp"
 	"strings"
 	"testing"
@@ -61,156 +59,156 @@ func Test_ParseUplink(t *testing.T) {
 	_ = date
 }
 
-func Test_ProcessFolder(t *testing.T) {
-	subDir := "rundowns_mix"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	opts := ArchiveOptions{
-		SourceDirectory:      srcDir,
-		DestinationDirectory: dstDir,
-		InvalidFileRename:    false,
-		// InvalidFileContinue:  false,
-		InvalidFileContinue:    true,
-		CompressionType:        "zip",
-		RecurseSourceDirectory: true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolder(t *testing.T) {
+// 	subDir := "rundowns_mix"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:      srcDir,
+// 		DestinationDirectory: dstDir,
+// 		InvalidFileRename:    false,
+// 		// InvalidFileContinue:  false,
+// 		InvalidFileContinue:    true,
+// 		CompressionType:        "zip",
+// 		RecurseSourceDirectory: true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
 
-func Test_ProcessFolderInvalid(t *testing.T) {
-	subDir := "rundowns_invalid"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	opts := ArchiveOptions{
-		SourceDirectory:        srcDir,
-		DestinationDirectory:   dstDir,
-		InvalidFileRename:      false,
-		InvalidFileContinue:    true,
-		CompressionType:        "zip",
-		RecurseSourceDirectory: true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolderInvalid(t *testing.T) {
+// 	subDir := "rundowns_invalid"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:        srcDir,
+// 		DestinationDirectory:   dstDir,
+// 		InvalidFileRename:      false,
+// 		InvalidFileContinue:    true,
+// 		CompressionType:        "zip",
+// 		RecurseSourceDirectory: true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
 
-func Test_ProcessFolderComplexNoDupes(t *testing.T) {
-	subDir := "rundowns_complex_nodupes"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	opts := ArchiveOptions{
-		SourceDirectory:          srcDir,
-		DestinationDirectory:     dstDir,
-		InvalidFileRename:        false,
-		InvalidFileContinue:      true,
-		CompressionType:          "zip",
-		PreserveFoldersInArchive: false,
-		RecurseSourceDirectory:   true,
-		// PreserveFoldersInArchive: true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolderComplexNoDupes(t *testing.T) {
+// 	subDir := "rundowns_complex_nodupes"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:          srcDir,
+// 		DestinationDirectory:     dstDir,
+// 		InvalidFileRename:        false,
+// 		InvalidFileContinue:      true,
+// 		CompressionType:          "zip",
+// 		PreserveFoldersInArchive: false,
+// 		RecurseSourceDirectory:   true,
+// 		// PreserveFoldersInArchive: true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
 
-func Test_ProcessFolderComplexDupes(t *testing.T) {
-	subDir := "rundowns_complex_dupes"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	opts := ArchiveOptions{
-		SourceDirectory:      srcDir,
-		DestinationDirectory: dstDir,
-		InvalidFileRename:    false,
-		// InvalidFileContinue:  false,
-		InvalidFileContinue:      true,
-		CompressionType:          "zip",
-		PreserveFoldersInArchive: false,
-		// RecurseSourceDirectory:   false,
-		RecurseSourceDirectory: true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolderComplexDupes(t *testing.T) {
+// 	subDir := "rundowns_complex_dupes"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:      srcDir,
+// 		DestinationDirectory: dstDir,
+// 		InvalidFileRename:    false,
+// 		// InvalidFileContinue:  false,
+// 		InvalidFileContinue:      true,
+// 		CompressionType:          "zip",
+// 		PreserveFoldersInArchive: false,
+// 		// RecurseSourceDirectory:   false,
+// 		RecurseSourceDirectory: true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
 
-func Test_ProcessFolderComplexDupesSame(t *testing.T) {
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, "rundowns_complex_dupes")
-	opts := ArchiveOptions{
-		SourceDirectory:          srcDir,
-		DestinationDirectory:     srcDir,
-		InvalidFileRename:        false,
-		InvalidFileContinue:      true,
-		CompressionType:          "zip",
-		PreserveFoldersInArchive: false,
-		RecurseSourceDirectory:   true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	fmt.Printf("%+v\n", process.Results)
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolderComplexDupesSame(t *testing.T) {
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, "rundowns_complex_dupes")
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:          srcDir,
+// 		DestinationDirectory:     srcDir,
+// 		InvalidFileRename:        false,
+// 		InvalidFileContinue:      true,
+// 		CompressionType:          "zip",
+// 		PreserveFoldersInArchive: false,
+// 		RecurseSourceDirectory:   true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	fmt.Printf("%+v\n", process.Results)
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
 
-func Test_ProcessFolderRundownsAppend(t *testing.T) {
-	subDir := "rundowns_append"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	subDirs := []string{
-		"dir1",
-		// "dir2",
-		// "dir3",
-		// "dir4",
-	}
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	for i := range subDirs {
-		srcSubDir := filepath.Join(srcDir, subDirs[i])
-		fmt.Println("PROCESSING FOLDER", srcSubDir)
-		opts := ArchiveOptions{
-			SourceDirectory:          srcSubDir,
-			DestinationDirectory:     dstDir,
-			InvalidFileRename:        false,
-			InvalidFileContinue:      true,
-			CompressionType:          "zip",
-			PreserveFoldersInArchive: false,
-			// PreserveFoldersInArchive: true,
-			ProcessedFileRename:    true,
-			RecurseSourceDirectory: true,
-		}
-		process := Archive{Options: opts}
-		err := process.Folder()
-		fmt.Printf("%+v\n", process.Results)
-		if err != nil {
-			t.Error(err)
-		}
-	}
-}
+// func Test_ProcessFolderRundownsAppend(t *testing.T) {
+// 	subDir := "rundowns_append"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	subDirs := []string{
+// 		"dir1",
+// 		// "dir2",
+// 		// "dir3",
+// 		// "dir4",
+// 	}
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	for i := range subDirs {
+// 		srcSubDir := filepath.Join(srcDir, subDirs[i])
+// 		fmt.Println("PROCESSING FOLDER", srcSubDir)
+// 		opts := ArchiveOptions{
+// 			SourceDirectory:          srcSubDir,
+// 			DestinationDirectory:     dstDir,
+// 			InvalidFileRename:        false,
+// 			InvalidFileContinue:      true,
+// 			CompressionType:          "zip",
+// 			PreserveFoldersInArchive: false,
+// 			// PreserveFoldersInArchive: true,
+// 			ProcessedFileRename:    true,
+// 			RecurseSourceDirectory: true,
+// 		}
+// 		process := Archive{Options: opts}
+// 		err := process.Folder()
+// 		fmt.Printf("%+v\n", process.Results)
+// 		if err != nil {
+// 			t.Error(err)
+// 		}
+// 	}
+// }
 
-func Test_ProcessFolderDate(t *testing.T) {
-	subDir := "rundowns_date"
-	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
-	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
-	opts := ArchiveOptions{
-		SourceDirectory:          srcDir,
-		DestinationDirectory:     dstDir,
-		InvalidFileRename:        false,
-		InvalidFileContinue:      true,
-		CompressionType:          "zip",
-		PreserveFoldersInArchive: false,
-		RecurseSourceDirectory:   true,
-	}
-	process := Archive{Options: opts}
-	err := process.Folder()
-	if err != nil {
-		t.Error(err)
-	}
-}
+// func Test_ProcessFolderDate(t *testing.T) {
+// 	subDir := "rundowns_date"
+// 	srcDir := filepath.Join(TEMP_DIR_TEST_SRC, subDir)
+// 	dstDir := filepath.Join(TEMP_DIR_TEST_DST, subDir)
+// 	opts := ArchiveOptions{
+// 		SourceDirectory:          srcDir,
+// 		DestinationDirectory:     dstDir,
+// 		InvalidFileRename:        false,
+// 		InvalidFileContinue:      true,
+// 		CompressionType:          "zip",
+// 		PreserveFoldersInArchive: false,
+// 		RecurseSourceDirectory:   true,
+// 	}
+// 	process := Archive{Options: opts}
+// 	err := process.Folder()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// }
