+++ internal/helper/debug_test.go
@@ -1,0 +1,11 @@
+package helper
+
+import (
+	"fmt"
+	"testing"
+)
+
+func Test_LogTraceFunction(t *testing.T) {
+	fmt.Println(TraceFunction(0))
+	fmt.Println(TraceFunction(1))
+}
