+++ internal/helper/errors_test.go
@@ -1,4 +1,4 @@
-package internal
+package helper
 
 import (
 	"fmt"
