+++ internal/helper/tester_test.go
@@ -1,39 +1,68 @@
 package helper
 
 import (
+	"fmt"
 	"log/slog"
 	"testing"
 )
 
+var testerConfig = TesterConfig{
+	TestDataSource: "../../test/testdata",
+}
+
 func TestMain(m *testing.M) {
-	Setup.InitMain()
-	exCode := m.Run()
-	slog.Warn("ex", "code", exCode)
-	Setup.waitGroup.Wait()
-	// Setup.CleanuUP()
-	// os.Exit(exCode)
+	testerConfig.InitMain()
+	exitCode := m.Run()
+	slog.Debug("exit code", "code", exitCode)
+	testerConfig.WaitGroup.Wait()
+	testerConfig.CleanuUP()
 }
 
-func TestABfail1(t *testing.T) {
-	defer Setup.RecoverPanic(t)
-	Setup.InitTest(t, true)
-	panic("kek")
-}
-
-func TestABfail2(t *testing.T) {
-	defer Setup.RecoverPanic(t)
-	Setup.InitTest(t, false)
-	panic("kek")
-}
-
-func TestABnofail1(t *testing.T) {
-	defer Setup.RecoverPanic(t)
-	Setup.InitTest(t, true)
+func TestABSimple(t *testing.T) {
+	kek := t.TempDir()
+	fmt.Println(kek)
 	Sleeper(2, "s")
 }
 
-func TestABnofail2(t *testing.T) {
-	defer Setup.RecoverPanic(t)
-	Setup.InitTest(t, false)
+func TestPABC1(t *testing.T) {
+	t.Parallel()
+	kek := t.TempDir()
+	fmt.Println(kek)
 	Sleeper(2, "s")
+}
+func TestPABC2(t *testing.T) {
+	t.Parallel()
+	kek := t.TempDir()
+	fmt.Println(kek)
+	Sleeper(2, "s")
+}
+
+func TestAABfail1(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, true)
+	panic("kek")
+}
+
+func TestAABfail2(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, false)
+	panic("kek")
+}
+
+func TestAABnofail1(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, true)
+	Sleeper(2, "s")
+}
+
+func TestAABnofail2(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, false)
+	Sleeper(2, "s")
+}
+
+func TestAAC1(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, true)
+	Sleeper(10, "s")
 }
