--- internal/testsetup_test.go
@@ -1,9 +1,0 @@
-package internal
-
-import "github/czech-radio/openmedia-archive/internal/helper"
-
-var Setup = helper.SetupTest{
-	TestDataSource:      "",
-	TempDataSource:      "",
-	TempDataDestination: "",
-}
