+++ internal/user/user_test.go
@@ -1,12 +1,32 @@
 package user
 
-// import (
-// 	// "github/czech-radio/openmedia-archive/internal/helper"
-// 	"github/czech-radio/openmedia-archive/internal/helper"
-// 	"testing"
-// )
+import (
+	// "github/czech-radio/openmedia-archive/internal/helper"
+	"fmt"
+	"github/czech-radio/openmedia-archive/internal/helper"
+	"log/slog"
+	"testing"
+)
 
-// func TestUserAB(t *testing.T) {
-// defer helper.Setup.RecoverPanic(t)
-// helper.Setup.InitTest(t, true)
-// }
+var testerConfig helper.TesterConfig
+
+func TestMain(m *testing.M) {
+	testerConfig.InitMain()
+	exitCode := m.Run()
+	slog.Debug("exit code", "code", exitCode)
+	testerConfig.WaitGroup.Wait()
+	testerConfig.CleanuUP()
+}
+
+func TestABSimple(t *testing.T) {
+	kek := t.TempDir()
+	fmt.Println(kek)
+	helper.Sleeper(2, "s")
+}
+
+func TestUserAAB(t *testing.T) {
+	defer testerConfig.RecoverPanic(t)
+	testerConfig.InitTest(t, true)
+	fmt.Println("kke")
+	helper.Sleeper(5, "s")
+}
